requirements.txt
numpy==1.26.4
mediapipe==0.10.9
opencv-python==4.9.0.80
rumps==0.4.0
pyobjc-framework-Quartz==10.1
PyYAML==6.0.1
